Add rendered cache for map sections and refactor W3MapSection
- Introduce LRUHashFast cache for managing rendered mesh instances
- Make W3MapSection movable, non-copyable with proper destructor
- Pass runtime manager to update/refresh methods
- Remove legacy LRUSetFast from surface header
- Update terrain, ground, and water rendering to use cache
- Adjust map node and test code for new API

+++ src/w3terrain/w3mapruntimemanager.h
@@ -21,8 +21,8 @@ public:
                 uint32_t uv_indices;  // four packed ground tileset texture indices
             };
             struct { // geo type
-                uint32_t tileset_id;
-                uint32_t geo_id;
+                uint16_t tileset_id;
+                uint16_t geo_id;
             };
         };
